refactor: move contract to accounts
